move coverage report path to account for github actions workflow
