--- .github/workflows/aws_ecr.yml
@@ -1,25 +1,0 @@
-name: AWS ECR #
-  # You may pin to the exact commit or the version.
-  # uses: kciter/aws-ecr-action@deed28ca73d748eba97d98847f903a3a77e702ad
-  uses: kciter/aws-ecr-action@v1
-  with:
-    # The AWS access key id
-    access_key_id: ${{ secrets.AWS_ACCESS_KEY_ID }}
-    # The AWS secret access key
-    secret_access_key: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
-    # AWS Account ID
-    account_id: 430746673.dkr.ecr.us-east-1.amazonaws.com/matt-oberlies-00000916-
-    # Name of your ECR repository
-    repo: rss-frontend
-    # The AWS region
-    region: us-east-1
-    # Set this to true to create the repository if it does not already exist
-    create_repo: # optional
-    # Comma-separated string of ECR image tags
-    tags: test # optional, default is latest
-    # Name of Dockerfile to use
-    dockerfile: # optional, default is Dockerfile
-    # Extra flags to pass to docker build (see docs.docker.com/engine/reference/commandline/build)
-    extra_build_args: # optional, default is 
-    # Path to Dockerfile, defaults to the working directory
-    path: ./rss-client # optional, default is .
