+++ .github/workflows/build.yml
@@ -27,8 +27,10 @@ jobs:
         run: |
           docker build -t $ECR_REGISTRY/$ECR_REPOSITORY:$IMAGE_TAG ./rss-client
           docker push $ECR_REGISTRY/$ECR_REPOSITORY:$IMAGE_TAG
+          docker tag $ECR_REGISTRY/$ECR_REPOSITORY:$IMAGE_TAG $ECR_REGISTRY/$ECR_REPOSITORY:latest
+          docker push $ECR_REGISTRY/$ECR_REPOSITORY:latest
   test:
-    name: download github repo to runner
+    name: Run Tests
     runs-on: ubuntu-latest
     steps:
       - name: Setup test environment
@@ -43,7 +45,7 @@ jobs:
           # Changing directories works only within the scope of the current step
           cd ./rss-client 
           npm ci
-          # npm run test
+          npm run test
       - name: Upload Code Coverage Report
         uses: actions/upload-artifact@v2
         with:
