+++ rss-client/src/app/Quiz/employee/cheater-warning/cheater-warning.component.spec.ts
@@ -1,7 +1,6 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { CheaterWarningComponent } from './cheater-warning.component';
-import { MatDialogModule } from '@angular/material/dialog';
 
 describe('CheaterWarningComponent', () => {
   let component: CheaterWarningComponent;
@@ -10,7 +9,7 @@ describe('CheaterWarningComponent', () => {
   beforeEach(async(() => {
     TestBed.configureTestingModule({
       declarations: [ CheaterWarningComponent ],
-      imports: [MatDialogModule]
+      imports: []
     })
     .compileComponents();
   }));
