+++ sonar-project.properties
@@ -7,6 +7,6 @@ sonar.language=ts
 sonar.tests=./rss-client
 sonar.test.inclusions=**/*.spec.ts
 
-sonar.javascript.lcov.reportPaths=.rss-client/coverage/rss-client/lcov.info
+sonar.javascript.lcov.reportPaths=./code-coverage-report/rss-client/lcov.info
 
 sonar.sourceEncoding=UTF-8
